authorized_keys.go: Provide Reload on authorized_keys
add `AuthorizedKeys.Reload`

This commit add a Reload method on AuthorizedKeys, allowing a user to reload keys
from an AuthorizedKeys file in a running application by calling the new
`Reload` method.

It is up to the user to implement how use it into an application. To give some example, it could come from an
API call or a SIGKILL which trigger a reload of the keys.

Signed-off-by: gaelL <contact@gaell.info>